Reshuffle module structure.
