+++ src/builder.rs
@@ -1,41 +1,9 @@
-//! Variable length octet sequences.
+//! Octets Builders
 //!
-//! This crate provides a set of basic traits that allow defining types that
-//! are generic over a variable length sequence of octets (or, vulgo: bytes).
-//! It implements these traits for most commonly used types of such sequences
-//! and provides a array-backed type for use in a no-std environment.
-//!
-//! # Types of Octet Sequences
-//!
-//! There are two fundamental types of octet sequences. If a sequence
-//! contains content of a constant size, we call it simply ‘octets.’ If the
-//! sequence is actually a buffer into which octets can be placed, it is
-//! called an `octets builder.`
-//!
-//! # Octets
-//!
-//! In their most simple form, any type that implements `AsRef<[u8]>` can
-//! serve as octets. However, in some cases additional functionality is
-//! required.
-//!
-//! The trait [`Octets`] allows taking a sub-sequence, called a ‘range’, out
-//! of the octets in the cheapest way possible. For most types, ranges will
-//! be octet slices `&[u8]` but some shareable types (most notably
-//! `bytes::Bytes`) allow ranges to be owned values, thus avoiding the
-//! lifetime limitations a slice would bring. Therefore, `Octets` allows
-//! defining the type of a range as an associated type.
-//!
-//! # Octets Builders
-//!
-//! Octets builders and their [`OctetsBuilder`] trait are comparatively
-//! straightforward. They represent a buffer to which octets can be appended.
+//! Octets builders, i.e., anything that implements the [`OctetsBuilder`]
+//! trait, represent a buffer to which octets can be appended.
 //! Whether the buffer can grow to accommodate appended data depends on the
-//! underlying type. Because it may not, all such operations may fail with an
-//! error defined by the trait implementation via
-//! `OctetsBuilder::AppendError`. Types where appending never fails (other
-//! than possibly panicking when running out of memory) should use the core
-//! library’s `Infallible` type here. This unlocks additional methods that
-//! don’t return a result and thus avoid an otherwise necessary unwrap.
+//! underlying type.
 //!
 //! The [`OctetsBuilder`] trait only provides methods to append data to the
 //! builder. Implementations may, however, provide more functionality. They
@@ -51,297 +19,11 @@
 //! * manipulation of already assembled octets requires `AsMut<[u8]>`, and
 //! * truncating the sequence of assembled octets happens through
 //!   [`Truncate`].
-//!
-//!
-//! # Conversion Traits
-//!
-//! A series of special traits allows converting octets into octets builder
-//! and vice versa. They pair octets with their natural builders via
-//! associated types. These conversions are always cyclic, i.e., if an
-//! octets value is converted into a builder and then that builder is
-//! converted back into an octets value, the initial and final octets value
-//! have the same type.
-//!
-//!
-//! # Using Trait Bounds
-//!
-//! When using these traits as bounds for generic types, always limit yourself
-//! to the most loose bounds you can get away with. Not all types holding
-//! octet sequences can actually implement all these traits, so by being too
-//! eager you may paint yourself into a corner.
-//!
-//! In many cases you can get away with a simple `AsRef<[u8]>` bound. Only use
-//! an explicit `Octets` bound when you need to return a range that may be
-//! kept around.
-//!
-//! Similarly, only demand of an octets builder what you actually need. Even
-//! something as seemingly trivial as `AsMut<[u8]>` isn’t a given. For
-//! instance, `Cow<[u8]>` doesn’t implement it but otherwise is a perfectly
-//! fine octets builder.
-
 
 use core::convert::Infallible;
-use core::ops::{Index, RangeBounds};
 #[cfg(feature = "bytes")] use bytes::{Bytes, BytesMut};
 #[cfg(feature = "std")] use std::borrow::Cow;
 #[cfg(feature = "std")] use std::vec::Vec;
-
-
-//============ Octets and Octet Builders =====================================
-
-//------------ Octets --------------------------------------------------------
-
-/// A type representing an octets sequence.
-///
-/// The primary purpose of the trait is to allow access to a sub-sequence,
-/// called a ‘range.’ The type of this range is given via the `Range`
-/// associated type. For most types it will be a `&[u8]` with a lifetime
-/// equal to that of a reference. Only if an owned range can be created
-/// cheaply, it should be that type.
-pub trait Octets: AsRef<[u8]> {
-    type Range<'a>: Octets where Self: 'a;
-
-    /// Returns a sub-sequence or ‘range’ of the sequence.
-    ///
-    /// # Panics
-    ///
-    /// The method should panic if `start` or `end` are greater than the
-    /// length of the octets sequence or if `start` is greater than `end`.
-    fn range(&self, range: impl RangeBounds<usize>) -> Self::Range<'_>;
-}
-
-impl<'t, T: Octets + ?Sized> Octets for &'t T {
-    type Range<'a> = <T as Octets>::Range<'a> where Self: 'a;
-
-    fn range(&self, range: impl RangeBounds<usize>) -> Self::Range<'_> {
-        (*self).range(range)
-    }
-}
-
-impl Octets for [u8] {
-    type Range<'a> = &'a [u8];
-
-    fn range(&self, range: impl RangeBounds<usize>) -> Self::Range<'_> {
-        self.index(
-            (range.start_bound().cloned(), range.end_bound().cloned())
-        )
-    }
-}
-
-#[cfg(feature = "std")]
-impl<'c> Octets for Cow<'c, [u8]> {
-    type Range<'a> = &'a [u8] where Self: 'a;
-
-    fn range(&self, range: impl RangeBounds<usize>) -> Self::Range<'_> {
-        self.as_ref().range(range)
-    }
-}
-
-#[cfg(feature = "std")]
-impl Octets for Vec<u8> {
-    type Range<'a> = &'a [u8];
-
-    fn range(&self, range: impl RangeBounds<usize>) -> Self::Range<'_> {
-        self.as_slice().range(range)
-    }
-}
-
-#[cfg(feature = "bytes")]
-impl Octets for Bytes {
-    type Range<'a> = Bytes;
-
-    fn range(&self, range: impl RangeBounds<usize>) -> Self::Range<'_> {
-        self.slice(range)
-    }
-}
-
-#[cfg(feature = "smallvec")]
-impl<A: smallvec::Array<Item = u8>> Octets for smallvec::SmallVec<A> {
-    type Range<'a> = &'a [u8] where A: 'a;
-
-    fn range(&self, range: impl RangeBounds<usize>) -> Self::Range<'_> {
-        self.as_slice().range(range)
-    }
-}
-
-
-//------------ Truncate ------------------------------------------------------
-
-/// An octet sequence that can be shortened.
-pub trait Truncate {
-    /// Truncate the sequence to `len` octets.
-    ///
-    /// If `len` is larger than the length of the sequence, nothing happens.
-    fn truncate(&mut self, len: usize);
-}
-
-impl<'a> Truncate for &'a [u8] {
-    fn truncate(&mut self, len: usize) {
-        if len < self.len() {
-            *self = &self[..len]
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl<'a> Truncate for Cow<'a, [u8]> {
-    fn truncate(&mut self, len: usize) {
-        match *self {
-            Cow::Borrowed(ref mut slice) => *slice = &slice[..len],
-            Cow::Owned(ref mut vec) => vec.truncate(len),
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl Truncate for Vec<u8> {
-    fn truncate(&mut self, len: usize) {
-        self.truncate(len)
-    }
-}
-
-#[cfg(feature = "bytes")]
-impl Truncate for Bytes {
-    fn truncate(&mut self, len: usize) {
-        self.truncate(len)
-    }
-}
-
-#[cfg(feature = "bytes")]
-impl Truncate for BytesMut {
-    fn truncate(&mut self, len: usize) {
-        self.truncate(len)
-    }
-}
-
-#[cfg(feature = "smallvec")]
-impl<A: smallvec::Array<Item = u8>> Truncate for smallvec::SmallVec<A> {
-    fn truncate(&mut self, len: usize) {
-        self.truncate(len)
-    }
-}
-
-//------------ OctetsFrom ----------------------------------------------------
-
-/// Convert a type from one octets type to another.
-///
-/// This trait allows creating a value of a type that is generic over an
-/// octets sequence from an identical value using a different type of octets
-/// sequence.
-///
-/// This is different from just `From` in that the conversion may fail if the
-/// source sequence is longer than the space available for the target type.
-pub trait OctetsFrom<Source>: Sized {
-    type Error;
-
-    /// Performs the conversion.
-    fn try_octets_from(source: Source) -> Result<Self, Self::Error>;
-
-    /// Performs an infallible conversion.
-    fn octets_from(source: Source) -> Self
-    where Self::Error: Into<Infallible> {
-        // XXX Use .into_ok() once that is stable.
-        match Self::try_octets_from(source) {
-            Ok(ok) => ok,
-            Err(_) => unreachable!()
-        }
-    }
-}
-
-impl<'a, Source: AsRef<[u8]> + 'a> OctetsFrom<&'a Source> for &'a [u8] {
-    type Error = Infallible;
-
-    fn try_octets_from(source: &'a Source) -> Result<Self, Self::Error> {
-        Ok(source.as_ref())
-    }
-}
-
-#[cfg(feature = "std")]
-impl<Source> OctetsFrom<Source> for Vec<u8>
-where
-    Self: From<Source>,
-{
-    type Error = Infallible;
-
-    fn try_octets_from(source: Source) -> Result<Self, Self::Error> {
-        Ok(From::from(source))
-    }
-}
-
-#[cfg(feature = "bytes")]
-impl<Source> OctetsFrom<Source> for Bytes
-where
-    Self: From<Source>,
-{
-    type Error = Infallible;
-
-    fn try_octets_from(source: Source) -> Result<Self, Self::Error> {
-        Ok(From::from(source))
-    }
-}
-
-#[cfg(feature = "bytes")]
-impl<Source> OctetsFrom<Source> for BytesMut
-where
-    Self: From<Source>,
-{
-    type Error = Infallible;
-
-    fn try_octets_from(source: Source) -> Result<Self, Self::Error> {
-        Ok(From::from(source))
-    }
-}
-
-#[cfg(features = "smallvec")]
-impl<Source, A> OctetsFrom<Source> for smallvec::SmallVec<A>
-where
-    Source: AsRef<u8>,
-    A: Array<Item = u8>,
-{
-    type Error = Infallible;
-
-    fn try_octets_from(source: Source) -> Result<Self, Self::Infallible> {
-        Ok(smallvec::ToSmallVec::to_smallvec(source.as_ref()))
-    }
-}
-
-
-//------------ OctetsInto ----------------------------------------------------
-
-/// Convert a type from one octets type to another.
-///
-/// This trait allows trading in a value of a type that is generic over an
-/// octets sequence for an identical value using a different type of octets
-/// sequence.
-///
-/// This is different from just `Into` in that the conversion may fail if the
-/// source sequence is longer than the space available for the target type.
-///
-/// This trait has a blanket implementation for all pairs of types where
-/// `OctetsFrom` has been implemented.
-pub trait OctetsInto<Target>: Sized {
-    type Error;
-
-    /// Performs the conversion.
-    fn try_octets_into(self) -> Result<Target, Self::Error>;
-
-    /// Performs an infallible conversion.
-    fn octets_into(self) -> Target
-    where Self::Error: Into<Infallible> {
-        match self.try_octets_into() {
-            Ok(ok) => ok,
-            Err(_) => unreachable!()
-        }
-    }
-}
-
-impl<Source, Target: OctetsFrom<Source>> OctetsInto<Target> for Source {
-    type Error = <Target as OctetsFrom<Source>>::Error;
-
-    fn try_octets_into(self) -> Result<Target, Self::Error> {
-        Target::try_octets_from(self)
-    }
-}
 
 
 //------------ OctetsBuilder -------------------------------------------------
@@ -729,12 +411,5 @@ impl<T, E, U: CollapseResult<T, E>> Collapse<U> for Result<T, E> {
     fn collapse(self) -> U {
         U::collapse_result(self)
     }
-}
-
-
-//============ Testing =======================================================
-
-#[cfg(test)]
-mod test {
 }
